Make tool act as a "Markdown browser"
So that an "offline wiki of markdown pages" can be constructed,
displayed, browsers, and edited.

Good for documentation / notes in a CM systems for consumption by the
team only, with out requiring team members to run a local web server,
without needing to transform MD to HTML on each build steps.

IE, when not using something like GIT HUB, but when using something like
TFS.